Updating

+++ ogmaneo/decoder.py
@@ -57,7 +57,7 @@ class Decoder:
                 area = diam * diam
                 num_weights = num_dendrites * hidden_size[3] * area * vld.size[2] * vld.size[3]
 
-                vl.weights = cl.clrandom.rand(cq, (num_weights,), np.float32, a=-0.01, b=0.01)
+                vl.weights = cl.clrandom.rand(cq, (num_weights,), np.uint8, a=-0.01, b=0.01)
                 vl.visible_states_prev = cl.array.zeros(cq, (num_visible_columns * vld.size[3],), np.int32)
 
                 self.vls.append(vl)
@@ -105,7 +105,7 @@ class Decoder:
                 area = diam * diam
                 num_weights = num_dendrites * self.hidden_size[3] * area * vld.size[2] * vld.size[3]
 
-                vl.weights = cl.array.empty(cq, (num_weights,), np.float32)
+                vl.weights = cl.array.empty(cq, (num_weights,), np.uint8)
                 vl.visible_states_prev = cl.array.empty(cq, (num_visible_columns * vld.size[3],), np.int32)
 
                 read_into_buffer(fd, vl.weights)
